sc2(runtime): avoid mounting devices for duplicated layers
If a container image has two layers with the same digest (likely because
the latter does not add any data modifications to the former), the nydus
snapshotter will generate one single disk image for the layer
(identified by its digest), but two different virtual volume devices.

If we pass this devices to the guest, the overlay module in the
guest-kernel will complain and fail to mount them.

In this patch we avoid passing duplicated devices, assuming that
duplications will always appear one after the other.